chore: move context modals types into separate file
