+++ context/modals/types.ts
@@ -1,0 +1,18 @@
+export type ModalType = 'ShareModal' | 'AuthModal';
+
+export type ShareModalState = {
+  isOpen: boolean;
+};
+
+export type AuthModalState = {
+  isOpen: boolean;
+};
+
+export type ModalState = {
+  shareModal: ShareModalState;
+  authModal: AuthModalState;
+};
+
+export type ModalsAction =
+  | { type: 'OPEN_MODAL'; modal: ModalType }
+  | { type: 'CLOSE_MODAL'; modal: ModalType };
